requirements.txt
requests
python-dateutil
pymongo
tornado
elasticsearch
markdown